search on type
